Move comments to separate files
